requirements.txt
Django>=1.10.0
djangorestframework>=1.9.0
pytest>=3.0.6
twine
pyjwt
cryptography
auth0-python
